requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
bcrypt
python-jose
python-multipart
pytest
requests
httpx